+++ molecularnodes/entities/ensemble/__init__.py
@@ -1,1 +1,3 @@
 from .ui import load_starfile, load_cellpack
+from .star import StarFile
+from .cellpack import CellPack
